+++ sync/sync_manager_test.go
@@ -192,7 +192,7 @@ func TestSync(t *testing.T) {
 		assert.NoError(t, err)
 		stakeTx := &transactions.StakeTransaction{
 			Validator_ID: valBytes,
-			Amount:       notes[0].Note.Amount,
+			Amount:       uint64(notes[0].Note.Amount),
 			Nullifier:    nullifier[:],
 			TxoRoot:      root[:],
 			Locktime:     0,
